be/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.23.2
